\name{subtractExpectation}

\alias{subtractExpectation}

\title{
  \code{subtractExpectation}
}

\description{
  subtractExpectation
}

\usage{
subtractExpectation(fg, bg, indTss, exppara, basal)
}

\arguments{
  \item{fg}{foreground}
  \item{bg}{backgroud}
  \item{indTss}{index of TSS}
  \item{exppara}{parameter for exppdf}
  \item{basal}{basal reads}
}

\details{
  One of the subtract functions.
}

\examples{
1:10
}

