\name{TssNorm-class}

\docType{class}

\alias{TssNorm}
\alias{TssNorm-class}
\alias{normalize}
\alias{normalize,TssData,function,integer,numeric,numeric,logical-method}

\title{
  Class \code{TssNorm}
}

\description{
  TssNorm
}

\section{Slots}{
  \describe{
    \item{\code{data}:}{data}
    \item{\code{region}:}{region}
    \item{\code{chr}:}{chr}
    \item{\code{date}:}{date}
    \item{\code{parameter}:}{parameter}
  }
}

\section{Methods}{
  \describe{
    \item{normalize:}{
      \code{obj="TssData", fun="function", offset="integer", basal="numeric",
        regpara="numeric", fit="logical"}
    }
  }
}

\keyword{classes}
\keyword{methods}
