\name{TssResult-class}

\docType{class}

\alias{TssResult}
\alias{TssResult-class}
\alias{identify}
\alias{identify,TssNorm,numeric,function-method}

\title{
  Class \code{TssResult}
}

\description{
  TssResult
}

\section{Slots}{
  \describe{
    \item{\code{data}:}{data}
    \item{\code{region}:}{region}
    \item{\code{chr}:}{chr}
    \item{\code{date}:}{date}
    \item{\code{parameter}:}{parameter}
  }
}

\section{Methods}{
  \describe{
    \item{identify:}{
      \code{obj="TssNorm", threshold="numeric", fun="function"}
    }
  }
}

\examples{
1:10
}

\keyword{classes}
\keyword{methods}
